ai-agents/requirements.txt
#!/usr/bin/env python3
# BunkerBuster AI Agent Swarm Requirements

# Core
python-dotenv==1.0.0
aiohttp==3.9.0
asyncio-contextmanager==1.0.0
requests==2.31.0

# Data Processing
pandas==2.0.0
numpy==1.24.0
scipy==1.10.0
scikit-learn==1.3.0

# NLP & LLM Integration
transformers==4.35.0
torch==2.1.0
openai==1.3.0
feedparser==6.0.10
tweepy==4.14.0
praw==7.7.0

# Geospatial
geopy==2.3.0
shapely==2.0.0
pyproj==3.6.0

# Signal Processing
librosa==0.10.0
scipy-signal==0.0.2

# Database
psycopg2-binary==2.9.0
sqlalchemy==2.0.0
asyncpg==0.29.0

# Message Queue & Cache
redis==5.0.0
pika==1.3.0

# Time Series
tslearn==0.6.3
statsmodels==0.14.0

# ML/Anomaly Detection
isolation-forest==0.1.3
pyod==1.1.0

# Utilities
python-dateutil==2.8.0
pytz==2023.3
pydantic==2.5.0
pyyaml==6.0
click==8.1.0
rich==13.7.0
tqdm==4.66.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Dev
jupyter==1.0.0
ipython==8.17.0
notebook==7.0.0
black==23.12.0
pylint==3.0.0
mypy==1.7.0
